add MDE environment client (#245)
Based on the internal model definition - tried plumbing this into the AWS SDK but since this is a local endpoint without the need for signers it didn't match up well (and their low level HTTP request methods are nested in a non exported util namespace). Instead I ended up pulling in got and throwing together this simple impl which given the simplicity of the API itself should suffice for the time being.