feat(CAWS): add telemetry (#555)
## Problem
* No named metrics
* Outdated CAWS model

## Solution
Add `caws_` metrics. The definitions were just copied from JB w/ some extra metadata added.

The Metric class is not intended to be final/complete. I'm still iterating on these ideas.
It's also unlikely that `telemetryName` will stick around for Commands. Coupling the easiest path for telemetry instrumentation to VS Code commands is not a good idea although it is convenient.

One important thing to point out is that this coupling to Commands made me want to pipe commands to other CAWS sub-modules. This is generally a bad sign because commands are meant to sit near the top of an architecture close to the UI.